Update PlayerController.cs
Changes;

- Added headers and grouped variables to make it easier to look at in inspector.
- Marked JumpPlayer() as obsolete.
- Marked isJumping as obsolete.
- Prevented the continuous jumps by detecting the frame when Jump was pressed.
- Changed the Jump() code to not need the isJumping variable.

I removed the isJumping variable by replacing it with checking if the player is in the air and if the jumpTimer has jump juice left. That way I can easily steal the remaining jump time if the player has stopped pressing jump or if he simply runs out of time.